+++ ereg/__init__.py
@@ -1,3 +1,4 @@
 # TODO do we need this?
+from .functional import registration_function
 from .registration import RegistrationClass
 from .utils.io import read_image_and_cast_to_32bit_float
